refactor: refactor after review
