+++ django_forest/utils/schema/__init__.py
@@ -1,98 +1,26 @@
 import copy
 import json
 import os
-import sys
 import logging
+from distutils.util import strtobool
+
 import django
 from django.conf import settings
 from django.db import connection
 from django.utils.module_loading import autodiscover_modules
 
-from django_forest.utils.apimap_errors import APIMAP_ERRORS
+from django_forest.utils.schema.apimap_errors import APIMAP_ERRORS
 from django_forest.utils.models import Models
 from django_forest.utils.get_type import get_type
 from django_forest.utils.json_api_serializer import create_json_api_schema
-from django_forest.utils.to_bool import to_bool
-
-if sys.version_info >= (3, 8):
-    from importlib import metadata
-else:
-    import importlib_metadata as metadata
-
 from django_forest.utils.forest_api_requester import ForestApiRequester
+from .definitions import COLLECTION, FIELD
+
+from .version import get_app_version
+
 
 # Get an instance of a logger
 logger = logging.getLogger(__name__)
-
-COLLECTION = {
-    'name': '',
-    'is_virtual': False,
-    'icon': None,
-    'is_read_only': False,
-    'is_searchable': True,
-    'only_for_relationships': False,
-    'pagination_type': 'page',
-    'search_fields': None,
-    'actions': [],
-    'segments': [],
-    'fields': []
-}
-
-FIELD = {
-    'field': '',
-    'type': '',
-    'is_filterable': True,
-    'is_sortable': True,
-    'is_read_only': False,
-    'is_required': False,
-    'is_virtual': False,
-    'default_value': None,
-    'integration': None,
-    'reference': None,
-    'inverse_of': None,
-    'relationship': None,
-    'widget': None,
-    'validations': []
-}
-
-ACTION = {
-    'name': '',
-    'type': 'bulk',
-    'baseUrl': None,
-    'endpoint': '',
-    'httpMethod': 'POST',
-    'redirect': None,
-    'download': False,
-    'fields': [],
-    'hooks': {
-        'load': False,
-        'change': []
-    }
-}
-
-ACTION_FIELD = {
-    'field': '',
-    'type': '',
-    'is_read_only': False,
-    'is_required': False,
-    'default_value': None,
-    'enums': None,
-    'integration': None,
-    'reference': None,
-    'description': None,
-    'widget': None,
-    'position': 0,
-}
-
-
-def get_app_version():
-    version = '0.0.0'
-    try:
-        version = metadata.version('django_forest')
-    except Exception:
-        pass
-    finally:
-        return version
 
 
 class Schema:
@@ -263,29 +191,36 @@ class Schema:
         }
 
     @staticmethod
-    def handle_apimap_error(status):
-        if APIMAP_ERRORS[status]['level'] == 'warning':
-            logger.warning(APIMAP_ERRORS[status]['message'])
-        else:
-            logger.error(APIMAP_ERRORS[status]['message'])
-
-    @classmethod
-    def _send_apimap(cls):
-        serialized_schema = cls.get_serialized_schema()
-        r = ForestApiRequester.post('forest/apimaps', serialized_schema)
+    def handle_status_code(r):
         if r.status_code in (200, 202, 204):
             r = r.json()
             if 'warning' in r:
                 logger.warning(r['warning'])
         elif r.status_code in APIMAP_ERRORS.keys():
-            cls.handle_apimap_error(r.status_code)
+            getattr(logger, APIMAP_ERRORS[r.status_code]['level'])(APIMAP_ERRORS[r.status_code]['message'])
         else:
-            logger.error(
-                'An error occured with the apimap sent to Forest. Please contact support@forestadmin.com for further investigations.')  # noqa E501
+            getattr(logger, APIMAP_ERRORS['error']['level'])(APIMAP_ERRORS['error']['message'])
+
+    @staticmethod
+    def get_disable_auto_schema_apply():
+        env_disable_auto_schema_apply = os.getenv('FOREST_DISABLE_AUTO_SCHEMA_APPLY', False)
+        disable_auto_schema_apply = getattr(settings, 'FOREST', {})\
+            .get('FOREST_DISABLE_AUTO_SCHEMA_APPLY', env_disable_auto_schema_apply)
+        if isinstance(disable_auto_schema_apply, str):
+            try:
+                disable_auto_schema_apply = strtobool(disable_auto_schema_apply)
+            except ValueError:
+                disable_auto_schema_apply = False
+        return disable_auto_schema_apply
 
     @classmethod
     def send_apimap(cls):
-        disable_auto_schema_apply = getattr(settings, 'FOREST', {}) \
-            .get('FOREST_DISABLE_AUTO_SCHEMA_APPLY', to_bool(os.getenv('FOREST_DISABLE_AUTO_SCHEMA_APPLY', False)))
+        disable_auto_schema_apply = cls.get_disable_auto_schema_apply()
         if not disable_auto_schema_apply:
-            cls._send_apimap()
+            serialized_schema = cls.get_serialized_schema()
+            try:
+                r = ForestApiRequester.post('forest/apimaps', serialized_schema)
+            except Exception:
+                logger.warning('Cannot send the apimap to Forest. Are you online?')
+            else:
+                cls.handle_status_code(r)
