--- django_forest/utils/to_bool.py
@@ -1,12 +1,0 @@
-TRUE_VALUES = {
-    't', 'T',
-    'y', 'Y', 'yes', 'YES',
-    'true', 'True', 'TRUE',
-    'on', 'On', 'ON',
-    '1', 1,
-    True
-}
-
-
-def to_bool(value):
-    return value in TRUE_VALUES
